Update README.md
The game of WAR is a classic card game that is typically played with a standard deck of 52 playing cards. The objective of the game is to beat the opponent by having a higher-ranking card in each round. The game is based on the concept of having a battle between two players, with each player playing a card from their deck and comparing the ranks of the cards. The player with the higher-ranking card wins the round and earns a point, while the losing card is added to a win pool. 

This game is a great way to teach children about strategy and competition, while also helping to improve their memory and cogitive skills. Also, the random chance element of the game adds a level of excitement and unpredictability, making it a fun and engaging experience for players of all ages.